Use main pings instead of saved_session
Saved_session pings are being deprecated.
However, main pings are both opt-in and opt-out, while saved_session
were just opt-in. For this reason we are filtering to include only
opt-in users, so the results should be similar.

Note that results will not be identical, since saved_session pings
often lag main pings, due to the main ping submission on date split.