Add ability to ignore certain clients
One client in particular was sending ~50k pings a day for an older
version of Aurora. This change reads a file from s3 with a listing
of clients to ignore and filters those pings out.

This change also refactorrd the aggregates code, bringing the filter
functions out of the aggregate_metrics function.