Add Cache Control and ETag headers
This change implements both cache control and ETag headers.

For Cache-Control:
For all requests but submission-date aggregates, the max-age
is set until the data is kicked from the local cache
(we know the response won't change until then).

For submission-date aggregates, max-age is always set to 24h.

ETags:
Etags are not set on any requests but submission-date aggregates.
The ETags are the same for all values, since submission-date
aggregates will never change, unless we do a backfill.

Thus, the single ETag value can be updated, invalidating all
previous ETags. This should only be done after a backfill.